Update style.css
The header title line height (site title) is different if you are on the home page or on another page. This is due to the fact that on the home page we have the title inside a H1. With this addition the header title will always have the same line height.